removes all warnings

+++ src/controller/mouse.h
@@ -10,19 +10,19 @@ namespace ctrl {
 
 namespace mouse {
 
-inline static void vertical_zoom_in(camera &cmr, size_t tick) noexcept {
+inline static void vertical_zoom_in(camera &cmr, double tick) noexcept {
   auto pos = cmr.lla();
   auto up = cmr.up();
 
-  pos.z -= pos.z * tick / 10;
+  pos.z -= static_cast<float>(pos.z * tick / 10.f);
   cmr.set_camera(pos, up);
 }
 
-inline static void vertical_zoom_out(camera &cmr, size_t tick) noexcept {
+inline static void vertical_zoom_out(camera &cmr, double tick) noexcept {
   auto pos = cmr.lla();
   auto up = cmr.up();
 
-  pos.z += pos.z * tick / 10;
+  pos.z += static_cast<float>(pos.z * tick / 10.f);
   cmr.set_camera(pos, up);
 }
